ADD: Se implementaron las excepciones recien creadas

+++ Backend/InstrumentsService/src/main/java/io/github/juniorcorzo/InstrumentsService/instruments/service/InstrumentsService.java
@@ -2,16 +2,17 @@ package io.github.juniorcorzo.InstrumentsService.instruments.service;
 
 import io.github.juniorcorzo.InstrumentsService.dto.ResponseWithData;
 import io.github.juniorcorzo.InstrumentsService.dto.ResponseWithoutData;
+import io.github.juniorcorzo.InstrumentsService.exception.FormatIdNotValid;
+import io.github.juniorcorzo.InstrumentsService.instruments.exceptions.InstrumentIdNotFound;
 import io.github.juniorcorzo.InstrumentsService.instruments.models.Instruments;
 import io.github.juniorcorzo.InstrumentsService.instruments.repositories.InstrumentsRepository;
 import io.github.juniorcorzo.InstrumentsService.utils.ResponseMessages;
 import lombok.AllArgsConstructor;
+import org.bson.types.ObjectId;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
 
 @Service
 @AllArgsConstructor
@@ -23,15 +24,19 @@ public class InstrumentsService {
                 HttpStatus.OK,
                 this.instrumentsRepository.findAll(),
                 ResponseMessages.OK.getMessage()
-                );
+        );
     }
 
     public ResponseWithData<Instruments> getById(String id) {
+        if (!ObjectId.isValid(id)) {
+            throw new FormatIdNotValid();
+        }
+
         return new ResponseWithData<>(
                 HttpStatus.OK,
                 Collections.singletonList(
                         this.instrumentsRepository.findById(id)
-                                .orElseThrow(() -> new RuntimeException("No existe un instrumento con ese id"))),
+                                .orElseThrow(InstrumentIdNotFound::new)),
                 ResponseMessages.OK.getMessage()
         );
     }
@@ -42,11 +47,17 @@ public class InstrumentsService {
     }
 
     public ResponseWithoutData updateInstruments(Instruments instruments) {
+        if (!ObjectId.isValid(instruments.getId())) throw new FormatIdNotValid();
+        if (!this.instrumentsRepository.existsById(instruments.getId())) throw new InstrumentIdNotFound();
+
         this.instrumentsRepository.save(instruments);
         return new ResponseWithoutData(HttpStatus.OK, ResponseMessages.OK.getMessage());
     }
 
     public ResponseWithoutData deleteInstruments(String id) {
+        if (!ObjectId.isValid(id)) throw new FormatIdNotValid();
+        if (!this.instrumentsRepository.existsById(id)) throw new InstrumentIdNotFound();
+
         this.instrumentsRepository.deleteById(id);
         return new ResponseWithoutData(HttpStatus.OK, ResponseMessages.OK.getMessage());
     }
